pyproject.toml
[tool.poetry]
name = "innohassle-roomsbot"
version = "0.1.0"
description = ""
authors = ["Maxim Fomin <m.fomin@innopolis.university>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3"
aiogram-dialog = "^2.1.0"
aiohttp = {version = "^3", extras = ["speedups"]}
pre-commit = "^4.0.1"
pydantic-settings = "^2"
python-dotenv = "^1.0.1"
requests = "^2.32.0"
ruff = "^0.6.9"
aiohttp-socks = "^0.10.2"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
